Add files via upload
Hello Everyone,
Today I am glad to share my one of dashboards Using Power Bi,
It is a TCS STOCK Analysis Dashboard. Here I will analyze the TCS NSE real-time Prices (in INR). I can Compare the opening and closing/ Lower Value and Higher Value of the stock by Year/Month and also do the chart to better understand the stock's performance. And I will use Year and Month as A filter, to filter the Performance of the stocks.